Make pythons RPCManager a context variable
