+++ changelog/pending/20230713--sdk-python--move-some-global-state-to-context-state-for-parallel-updates.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: fix
+  scope: sdk/python
+  description: Move some global state to context state for parallel updates.
