+++ sdk/python/lib/pulumi/automation/_server.py
@@ -17,14 +17,14 @@ import logging
 import sys
 import traceback
 from contextlib import suppress
+
 import grpc
 
-from ._workspace import PulumiFn
 from .. import log
-from ..runtime.proto import language_pb2, plugin_pb2, LanguageRuntimeServicer
-from ..runtime import run_in_stack, reset_options, set_all_config
-from ..runtime.rpc_manager import RPC_MANAGER
 from ..errors import RunError
+from ..runtime import reset_options, run_in_stack, set_all_config
+from ..runtime.proto import LanguageRuntimeServicer, language_pb2, plugin_pb2
+from ._workspace import PulumiFn
 
 _py_version_less_than_3_7 = sys.version_info[0] == 3 and sys.version_info[1] < 7
 
@@ -122,7 +122,6 @@ class LanguageServer(LanguageRuntimeServicer):
             loop.close()
             sys.stdout.flush()
             sys.stderr.flush()
-            RPC_MANAGER.clear()
 
         return result
