+++ sdk/python/lib/pulumi/runtime/rpc.py
@@ -16,45 +16,46 @@ Support for serializing and deserializing properties going into or flowing
 out of RPC calls.
 """
 import asyncio
-from collections import abc
 import functools
 import inspect
 from abc import ABC, abstractmethod
+from collections import abc
+from enum import Enum
 from typing import (
-    List,
+    TYPE_CHECKING,
     Any,
     Callable,
     Dict,
+    Iterable,
+    List,
     Mapping,
     Optional,
     Sequence,
     Set,
     Union,
-    TYPE_CHECKING,
     cast,
 )
-from enum import Enum
 
+import six
 from google.protobuf import struct_pb2
 from semver import VersionInfo as Version
-import six
-from . import known_types, settings
-from .resource import _expand_dependencies
-from .. import log
-from .. import _types
+
+from .. import _types, log
 from .. import urn as urn_util
+from . import known_types, settings
+from .resource_cycle_breaker import declare_dependency
 
 if TYPE_CHECKING:
-    from ..output import Inputs, Input, Output
-    from ..resource import Resource, CustomResource, ProviderResource
     from ..asset import (
+        AssetArchive,
+        FileArchive,
         FileAsset,
+        RemoteArchive,
         RemoteAsset,
         StringAsset,
-        FileArchive,
-        RemoteArchive,
-        AssetArchive,
     )
+    from ..output import Input, Inputs, Output
+    from ..resource import CustomResource, ProviderResource, Resource
 
 UNKNOWN = "04da6b54-80e4-46f7-96ec-b56ff0331ba9"
 """If a value is None, we serialize as UNKNOWN, which tells the engine that it may be computed later."""
@@ -224,6 +225,78 @@ async def serialize_properties(
             property_deps[translated_name] = deps
 
     return struct
+
+
+async def _add_dependency(
+    deps: Set[str], res: "Resource", from_resource: Optional["Resource"]
+):
+    """
+    _add_dependency adds a dependency on the given resource to the set of deps.
+
+    The behavior of this method depends on whether or not the resource is a custom resource, a local component resource,
+    or a remote component resource:
+
+    - Custom resources are added directly to the set, as they are "real" nodes in the dependency graph.
+    - Local component resources act as aggregations of their descendents. Rather than adding the component resource
+      itself, each child resource is added as a dependency.
+    - Remote component resources are added directly to the set, as they naturally act as aggregations of their children
+      with respect to dependencies: the construction of a remote component always waits on the construction of its
+      children.
+
+    In other words, if we had:
+
+                     Comp1
+                 |     |     |
+             Cust1   Comp2  Remote1
+                     |   |       |
+                 Cust2   Cust3  Comp3
+                 |                 |
+             Cust4                Cust5
+
+    Then the transitively reachable resources of Comp1 will be [Cust1, Cust2, Cust3, Remote1].
+    It will *not* include:
+    * Cust4 because it is a child of a custom resource
+    * Comp2 because it is a non-remote component resoruce
+    * Comp3 and Cust5 because Comp3 is a child of a remote component resource
+    """
+
+    # Exit early if there are cycles to avoid hangs.
+    no_cycles = declare_dependency(from_resource, res) if from_resource else True
+    if not no_cycles:
+        return
+
+    from .. import ComponentResource  # pylint: disable=import-outside-toplevel
+
+    # Local component resources act as aggregations of their descendents.
+    # Rather than adding the component resource itself, each child resource
+    # is added as a dependency.
+    if isinstance(res, ComponentResource) and not res._remote:
+        # Copy the set before iterating so that any concurrent child additions during
+        # the dependency computation (which is async, so can be interleaved with other
+        # operations including child resource construction which adds children to this
+        # resource) do not trigger modification during iteration errors.
+        child_resources = res._childResources.copy()
+        for child in child_resources:
+            await _add_dependency(deps, child, from_resource)
+        return
+
+    urn = await res.urn.future()
+    if urn:
+        deps.add(urn)
+
+
+async def _expand_dependencies(
+    deps: Iterable["Resource"], from_resource: Optional["Resource"]
+) -> Set[str]:
+    """
+    _expand_dependencies expands the given iterable of Resources into a set of URNs.
+    """
+
+    urns: Set[str] = set()
+    for d in deps:
+        await _add_dependency(urns, d, from_resource)
+
+    return urns
 
 
 # pylint: disable=too-many-return-statements, too-many-branches
@@ -508,12 +581,12 @@ def deserialize_properties(
     # which has type `Struct` in our gRPC proto definition.
     if _special_sig_key in props_struct:
         from .. import (  # pylint: disable=import-outside-toplevel
-            FileAsset,
-            StringAsset,
-            RemoteAsset,
             AssetArchive,
             FileArchive,
+            FileAsset,
             RemoteArchive,
+            RemoteAsset,
+            StringAsset,
         )
 
         if props_struct[_special_sig_key] == _special_asset_sig:
